# public/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: isalesbookv2 API
  version: 1.0.0
  description: |
    API and documentation for isalesbookv2 modular monolith (Email, Logger, Registration)
  contact:
    email: admin@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local API server
  - url: http://backend-v2.isalesbook.com
    description: Production API server
paths:
...
  /api/registration/register:
    post:
      summary: Register a new user
      description: |
        Register a new user (individual or company). After successful registration, a verification email is sent to the user containing their organization code and verification link. The user must verify their email to activate their account.
      tags: [Registration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - title: Individual Registration
                  type: object
                  required:
                    - first_name
                    - last_name
                    - email
                    - phone_number
                    - company_address
                    - company_url
                    - country_id
                    - state_id
                    - company_name
                    - password
                    - password_confirmation
                    - type
                  properties:
                    first_name:
                      type: string
                    last_name:
                      type: string
                    email:
                      type: string
                      format: email
                    phone_number:
                      type: string
                    company_address:
                      type: string
                    company_url:
                      type: string
                    country_id:
                      type: integer
                    state_id:
                      type: integer
                    company_name:
                      type: string
                    password:
                      type: string
                      minLength: 8
                    password_confirmation:
                      type: string
                    type:
                      type: integer
                      enum: [1]
                - title: Company Registration
                  type: object
                  required:
                    - company_name
                    - company_contact_person
                    - company_contact_number
                    - email
                    - company_address
                    - company_url
                    - country_id
                    - state_id
                    - password
                    - password_confirmation
                    - type
                  properties:
                    company_name:
                      type: string
                    company_contact_person:
                      type: string
                    company_contact_number:
                      type: string
                    email:
                      type: string
                      format: email
                    company_address:
                      type: string
                    company_url:
                      type: string
                    country_id:
                      type: integer
                    state_id:
                      type: integer
                    password:
                      type: string
                      minLength: 8
                    password_confirmation:
                      type: string
                    type:
                      type: integer
                      enum: [2]
      responses:
        '201':
          description: User registered. A verification email is sent to the user with their organization code and verification link.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                  message:
                    type: string
                    example: "Registration successful. Please check your email for a verification link and your organization code."
  /api/registration/verify-email:
    post:
      summary: Verify email with token
      description: |
        Verify a user's email address using the token sent in the verification email. The verification email contains the user's organization code and a verification link. Required for account activation.
      tags: [Registration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid or expired token
  /api/registration/resend-verification:
    post:
      summary: Resend verification email
      description: |
        Resend the verification email to the user. The email contains the organization code and verification link.
      tags: [Registration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Verification email resent
          content:
            application/json:
              schema:
                type: object
  /api/registration/forgot-password:
    post:
      summary: Send password reset email
      description: |
        Send a password reset email to the user. The email is styled and contains a secure reset link.
      tags: [Registration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                type: object
  /api/registration/reset-password:
    post:
      summary: Reset password
      tags: [Registration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password, password_confirmation]
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 8
                password_confirmation:
                  type: string
      responses:
        '200':
          description: Password reset
          content:
            application/json:
              schema:
                type: object
  /api/registration/check-availability:
    get:
      summary: Check availability of email, username, or phone
      tags: [Registration]
      parameters:
        - in: query
          name: email
          schema:
            type: string
            format: email
        - in: query
          name: username
          schema:
            type: string
        - in: query
          name: phone
          schema:
            type: string
      responses:
        '200':
          description: Availability checked
          content:
            application/json:
              schema:
                type: object
...
  /api/registration/profile:
    get:
      summary: Get user profile
      tags: [Registration]
      responses:
        '200':
          description: Profile retrieved
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
    put:
      summary: Update user profile
      tags: [Registration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                username:
                  type: string
                phone:
                  type: string
                date_of_birth:
                  type: string
                  format: date
                gender:
                  type: string
                  enum: [male, female, other, prefer_not_to_say]
                bio:
                  type: string
                website:
                  type: string
                  format: url
                location:
                  type: string
                timezone:
                  type: string
                language:
                  type: string
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
        '422':
          description: Validation error
  /api/registration/change-password:
    post:
      summary: Change user password
      tags: [Registration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password, new_password_confirmation]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
                  minLength: 8
                new_password_confirmation:
                  type: string
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                type: object
        '422':
          description: Validation error
  /api/registration/users:
    get:
      summary: Search users (Admin only)
      tags: [Registration]
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive]
        - in: query
          name: verified
          schema:
            type: string
            enum: [yes, no]
        - in: query
          name: per_page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Users retrieved
          content:
            application/json:
              schema:
                type: object 